+++ NEWS.md
@@ -1,5 +1,19 @@
 # mlr3 (development version)
 
+## New Features:
+
+* `Task` got method `$materialize_view()` which can save memory after subsetting a task.
+* Better input validation for:
+  * `Learner` fields.
+* Various improvements to the documentation and logging output, including
+  examples for methods.
+* Measure "oob_error" now works even without storing models during resampling.
+
+## Deprecations:
+
+* Assigning to some fields of `Task`, `Learner`, and `Resampling` now throws a deprecation warning.
+  This will become an error in the future.
+
 # mlr3 1.1.0
 
 * feat: Add new measure `MeasureRegrRQR` for quantile regression.